src/Infra/Gateways/awsCognitoIdentityProvider/actions/deleteUser.ts
/* eslint-disable @typescript-eslint/no-non-null-assertion */
import { AwsCognitoTemplate } from '../../Templates/AWS';
import { IDeleteUser } from '../../../Interfaces/Gateways';
import { AwsCognitoTemplateConstructorParams } from './Types';

type ExecuteInput = IDeleteUser.Input

export class DeleteUser extends AwsCognitoTemplate {
  constructor({ clientId, cognitoInstance }: AwsCognitoTemplateConstructorParams) {
    super({ clientId, cognitoInstance });
  }

  protected async performAction({ accessToken }: ExecuteInput): Promise<void> {
    await this.serviceInstance.deleteUser({
      AccessToken: accessToken,
    }).promise();
  }
}
